requirements.txt
# Apache Airflow (core)
apache-airflow==2.8.1

# Airflow Providers (essential)
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-http==4.6.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# ML (Phase 4)
scikit-learn==1.3.2
xgboost==1.7.6

# dbt (Phase 3)
dbt-core==1.11.0b3
dbt-bigquery==1.10.2

# Data quality (Phase 5)
great-expectations==0.18.5

# BigQuery (Phase 3)
google-cloud-bigquery==3.40.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
